# openapi.yml
openapi: '3.1.0'
info:
  title: Comprehensive Books Management REST API
  description: >-
    **A REST API for managing Users, Books, Authors, Genres, and their interactions, developed with [Express](https://expressjs.com).**

    This API provides endpoints to manage Users, Books, Authors, Genres, and the relationships between them.
  version: 1.0.0
  contact:
    url: https://github.com/Broannn/BookNest-RestAPI
servers:
  - url: https://booknest-restapi.onrender.com/
...
paths:
  /users:
    post:
      summary: Create a user.
      tags:
        - users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserData'
      responses:
        201:
          description: The user was successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        422:
          description: Validation error.
    get:
      summary: List users.
      tags:
        - users
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
      responses:
        200:
          description: A list of users.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /users/{userId}: # Corrected indentation
    get:
      summary: Get a user by ID.
      tags:
        - users
      parameters:
        - name: userId
          in: path
          required: true
          description: The ID of the user.
          schema:
            type: string
      responses:
        200:
          description: The user details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        404:
          description: User not found.
    put:
      summary: Update a user.
      tags:
        - users
      parameters:
        - name: userId
          in: path
          required: true
          description: The ID of the user.
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserData'
      responses:
        200:
          description: User successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        404:
          description: User not found.
    delete:
      summary: Delete a user.
      tags:
        - users
      parameters:
        - name: userId
          in: path
          required: true
          description: The ID of the user.
          schema:
            type: string
      responses:
        204:
          description: User successfully deleted.
        404:
          description: User not found.
  /users/{userId}/favorites:
    post:
      summary: Add a book to user's favorites.
      tags:
        - interactions
      parameters:
        - name: userId
          in: path
          required: true
          description: The ID of the user.
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                bookId:
                  type: string
                  description: The ID of the book to be added.
      responses:
        201:
          description: Book added to favorites.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Favorite'
    get:
      summary: Get user's favorite books.
      tags:
        - interactions
      parameters:
        - name: userId
          in: path
          required: true
          description: The ID of the user.
          schema:
            type: string
      responses:
        200:
          description: A list of favorite books.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
    delete:
      summary: Remove a book from user's favorites.
      tags:
        - interactions
      parameters:
        - name: userId
          in: path
          required: true
          description: The ID of the user.
          schema:
            type: string
        - name: bookId
          in: path
          required: true
          description: The ID of the book to be removed.
          schema:
            type: string
      responses:
        204:
          description: Book successfully removed from favorites.
  /users/{userId}/wishlist:
    post:
      summary: Add a book to user's wishlist.
      tags:
        - interactions
      parameters:
        - name: userId
          in: path
          required: true
          description: The ID of the user.
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                bookId:
                  type: string
                  description: The ID of the book to be added.
      responses:
        201:
          description: Book added to wishlist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wishlist'
    get:
      summary: Get user's wishlist.
      tags:
        - interactions
      parameters:
        - name: userId
          in: path
          required: true
          description: The ID of the user.
          schema:
            type: string
      responses:
        200:
          description: A list of books in the wishlist.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
    delete:
      summary: Remove a book from user's wishlist.
      tags:
        - interactions
      parameters:
        - name: userId
          in: path
          required: true
          description: The ID of the user.
          schema:
            type: string
        - name: bookId
          in: path
          required: true
          description: The ID of the book to be removed.
          schema:
            type: string
      responses:
        204:
          description: Book successfully removed from wishlist.
...
components:
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
      example: 1
    pageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
      example: 25
...
  schemas:
    Genre: # Fixed indentation for Genre schema
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the genre.
          example: 5fbd827a512e73129cfedb1c
        name:
          type: string
          description: The name of the genre.
          example: Fiction
        createdAt:
          type: string
          format: date-time
          description: The date the genre was created.
          example: 2025-01-13T10:00:00.000Z
      required:
        - id
        - name
...
    User:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the user.
          example: 7fe91a9a41bc839033eedf0f
        username:
          type: string
          description: The username of the user.
          example: johndoe
        email:
          type: string
          description: The email of the user.
          example: johndoe@example.com
      required:
        - id
        - username
        - email
    Favorite:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the favorite.
          example: 123abc
        userId:
          type: string
          description: The ID of the user who favorited the book.
          example: 7fe91a9a41bc839033eedf0f
        bookId:
          type: string
          description: The ID of the favorited book.
          example: 123bookid
      required:
        - id
        - userId
        - bookId
    Wishlist:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the wishlist item.
          example: 456wishlistid
        userId:
          type: string
          description: The ID of the user who added the book to the wishlist.
          example: 7fe91a9a41bc839033eedf0f
        bookId:
          type: string
          description: The ID of the wishlisted book.
          example: 789bookid
      required:
        - id
        - userId
        - bookId
    Book: # Corrected indentation starts here
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the book.
          example: 123bookid
        title:
          type: string
          description: The title of the book.
          example: Great Expectations
        authorId:
          type: string
          description: The ID of the author of the book.
          example: 7a9bc839033eedf0f
        genres:
          type: array
          description: Genres associated with the book.
          items:
            $ref: '#/components/schemas/Genre'
        publicationDate:
          type: string
          format: date
          description: The publication date of the book.
          example: 2023-01-01
      required:
        - id
        - title
        - authorId
...